Move Used table unit tests to unit tests module
